requirements.txt
asgiref==3.7.2
autopep8==2.0.4
Django==5.0.2
django-cors-headers==4.3.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
pycodestyle==2.11.1
PyJWT==2.8.0
pytz==2024.1
sqlparse==0.4.4
